+++ src/main/webapp/wise5/components/html/index.html
@@ -1,62 +1,3 @@
 <div ng-controller='HTMLController as htmlController'>
-  <div ng-if='htmlController.mode === "authoring"'>
-    <div class='advancedAuthoringDiv'
-       ng-if='htmlController.showAdvancedAuthoring'>
-      <div>
-        <h6>{{ 'advancedAuthoringOptions' | translate }}</h6>
-      </div>
-      <div>
-        <md-input-container style='margin-right: 20px; width: 150px; height: 25px;'>
-          <label>{{ 'COMPONENT_WIDTH' | translate }}</label>
-          <input type='number'
-               ng-model='htmlController.authoringComponentContent.componentWidth'
-               ng-model-options='{ debounce: 1000 }'
-               ng-change='htmlController.authoringViewComponentChanged()'/>
-        </md-input-container>
-      </div>
-      <div>
-        <md-button class='topButton md-raised md-primary' ng-click='htmlController.showRubricAuthoring = !htmlController.showRubricAuthoring'>
-          <md-icon>message</md-icon>
-          <md-tooltip md-direction='top' class='projectButtonTooltip'>{{ 'EDIT_COMPONENT_RUBRIC' | translate }}</md-tooltip>
-        </md-button>
-        <div ng-if='htmlController.showRubricAuthoring'>
-          <summernote id='{{htmlController.summernoteRubricId}}'
-                ng-model='htmlController.summernoteRubricHTML'
-                ng-change='htmlController.summernoteRubricHTMLChanged()'
-                config='htmlController.summernoteRubricOptions'
-                ng-model-options='{ debounce: 1000 }'
-                rows='10'
-                cols='100'>
-          </summernote>
-        </div>
-      </div>
-      <div>
-        <label class='node__label--vertical-alignment'>{{ 'json' | translate }}</label>
-        <md-button class='topButton md-raised md-primary'
-               ng-click='htmlController.showJSONButtonClicked()'>
-          <md-icon>code</md-icon>
-          <md-tooltip md-direction='top' class='projectButtonTooltip'>{{ 'SHOW_JSON' | translate }}</md-tooltip>
-        </md-button>
-        <span ng-if='htmlController.jsonStringChanged'>
-          {{ 'closeTheJSONViewToSaveTheChanges' | translate }}
-        </span>
-        <div ng-if='htmlController.showJSONAuthoring'>
-          <md-input-container style='width: 100%;'>
-            <textarea ng-model='htmlController.authoringComponentContentJSONString'
-                  style='width: 90%; border: 1px solid black;'
-                  ng-change='htmlController.authoringJSONChanged()'></textarea>
-          </md-input-container>
-        </div>
-      </div>
-    </div>
-    <br/>
-    <summernote id='{{htmlController.summernotePromptId}}'
-          ng-model='htmlController.summernotePromptHTML'
-          ng-change='htmlController.summernotePromptHTMLChanged()'
-          config='htmlController.summernotePromptOptions'
-          ng-model-options='{ debounce: 1000 }'
-          rows='10'
-          cols='100'></summernote>
-  </div>
   <div ng-if='htmlController.mode !== "authoring"'><compile data='htmlController.componentContent.html'></compile></div>
 </div>
