+++ src/main/webapp/wise5/directives/component/component.js
@@ -84,9 +84,14 @@ var ComponentController = function ComponentController($injector, $scope, $compi
         $scope.nodeAuthoringController = $scope.$parent.nodeAuthoringController;
     }
 
+    if ($scope.mode === 'authoring' && componentContent.type == 'HTML') {
+        $scope.componentTemplatePath = this.NodeService.getComponentAuthoringTemplatePath(componentContent.type);
+    } else {
+        $scope.componentTemplatePath = this.NodeService.getComponentTemplatePath(componentContent.type);
+    }
+
     $scope.componentContent = componentContent;
     $scope.componentState = this.componentState;
-    $scope.componentTemplatePath = this.NodeService.getComponentTemplatePath(componentContent.type);
     $scope.nodeId = this.nodeId;
     $scope.workgroupId = this.workgroupId;
     $scope.teacherWorkgroupId = this.teacherWorkgroupId;
